pyproject.toml
[project]
# Update the project name to match your actual project name
name = "taxi_dlt_pkg"
version = "0.1.0"
# Update with your email or organization name for your environment
authors = [{ name = "your-email@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "pyspark>=3.5.0",  # For local Spark testing
    "delta-spark>=3.0.0",  # For Delta Lake support in local tests
]

[project.scripts]
# Entry point removed - this is a pure transformation package
# No CLI scripts needed for DLT pipeline transformations

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 125
